requirements.txt
streamlit==1.45.1
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.25
langchain-huggingface==0.3.0
openai==1.86.0
qdrant-client==1.14.3
pydantic==2.11.7
python-dotenv==1.0.1
einops==0.8.1
langgraph==0.4.9
rapidfuzz==3.13.0
requests==2.32.3
